存放位置变更列表、变更存储位置

+++ src/main/java/com/pkusoft/agxt/model/CabSpace.java
@@ -83,7 +83,7 @@ public class CabSpace {
      */
     @Column(name = "SN")
     @ApiModelProperty(value = "排序序号")
-    private Integer sn;
+    private Double sn;
 
  	/**
      * 所有人ID
@@ -451,7 +451,7 @@ public class CabSpace {
     * 获得排序序号
     * @return SN
     */
-    public Integer getSn() {
+    public Double getSn() {
         return this.sn;
     }
 
@@ -459,7 +459,7 @@ public class CabSpace {
     * 设置排序序号
     * @param sn
     */
-    public void setSn(Integer sn) {
+    public void setSn(Double sn) {
         this.sn = sn;
     }
